create a providers component
