+++ src/app/layout.tsx
@@ -1,13 +1,9 @@
 import type { Metadata } from "next";
 import { Roboto } from "next/font/google";
 import "./globals.css";
-import { getServerSession } from "next-auth";
-import SessionProvider from "@/components/SessionProvider";
-import AuthKitProvider from '@/components/AuthKitProvider';
 import Header from "@/components/Header";
 import Footer from "@/components/Footer";
-import { DOMAIN } from "@/utils/config";
-import NextUIProvider from '@/components/NextUIProvider';
+import Providers from "@/components/Providers";
 
 const font = Roboto({ subsets: ["latin"], weight: '400' });
 
@@ -16,25 +12,16 @@ export const metadata: Metadata = {
   description: "Look who is online and onchain.",
 };
 
-const config = {
-  relay: 'https://relay.farcaster.xyz',
-  rpcUrl: 'https://mainnet.optimism.io',
-  domain: DOMAIN,
-  siweUri: `http://${DOMAIN}/`,
-};
-
-export default async function RootLayout({
+export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
 
-  const session = await getServerSession();
-
   return (
-    <html lang="en">
-      <body className={`${font.className} box-border pt-20 flex flex-col`} style={{backgroundColor: '#313131'}}>
-        <SessionProvider session={session}><AuthKitProvider config={config}><NextUIProvider>
+    <html lang="en" style={{backgroundColor: '#313131'}}>
+      <body className={`${font.className} box-border pt-20 flex flex-col bg-inherit`}>
+        <Providers>
 
           {/* Header */}
           <Header />
@@ -45,7 +32,7 @@ export default async function RootLayout({
           {/* Footer */}
           <Footer />
 
-          </NextUIProvider></AuthKitProvider></SessionProvider>
+        </Providers>
       </body>
     </html>
   );
